AppOps: change MODE_ASK to ALLOWED for receiving sms
SMS is received via broadcast. The operation checking doesn't
popup dialog to users because of broadcast ANR issue. So if the
mode is MODE_ASK and there is no popup dialog, then sms will be
ignored forever.

Meanwhile, receiving sms operation check is not the requirement
of appops. So one solution is to modify default mode for this
operation from MODE_ASK to ALLOWED and doesn't give user the
chance to modify it.

Change-Id: I2b6f90f13d3b2576bcbc6587ab2c6efa47c8dd52
Signed-off-by: houst0nn <houstonn@gmail.com>