SystemUI: Stick Pie to edge on phones
Apply the same UX of AOSP's navbar on phones to Pie
controls. When rotating the phone, Pie now sticks to the
same physical screen edge, to which it was pinned before. As
most people start holding their phone with two hands in
landscape mode, keeping Pie at the right or left edge might
be more comfortable to reach by thumbs. The behavior on
tablets should not be affected in any way by this change.

Changes:

PieControlPanel stores only the absolute gravity (the
rotation independent position in portrait mode) of Pie.
PieMenu and PieController query the gravity from Panel.
Query methods (getOrientation(), getDegree(),
isGravityPossible()) only return the relative gravity (the
position depending on current rotation). Therfore, only the
PieControlPanel is aware of, whether and how gravity is
converted.

The amount of snap points is reduced in landscape mode,
because Pie cannot be pinned to the top edge of the screen.

Change-Id: I263f16e1c59e134f374b9a3a007ecdce2d54dc46

Conflicts:
	packages/SystemUI/res/values-sw600dp/config.xml