Keyguard: Avoid unneeded SIM dialog show() calls
https://paranoidandroid.atlassian.net/browse/AOSPA-465

This commit makes the progress dialog's show() get called only once
each time the SIM unlock process is started.

It was reported that a phone can be frozen by entering the PIN code
for the SIM card "too quickly". This leaves the user with a device on
which the screen is filled with a non-cancellable dialog box saying
that the SIM card is being unlocked. The logs inform that Keyguard
experienced a success and no explicit errors could be seen there and
only the dialog stayed around, all leading to believe the issue is
with how the progress dialog is being showed and later hidden (or
simply dismissed), in which case this show() call reduction might be
benefitial.

Change-Id: Ibbb4b725598c95e2b2ee9689f6bf50247ad235ac

Conflicts:
	packages/Keyguard/src/com/android/keyguard/KeyguardSimPukView.java