Pie: Fix Pie positioning when PIE_ALWAYS_AT_RIGHT
Locking Pie to a single physical side of the screen makes perfect
sense when it's done on the short side of the screen (bottom in
portrait) as then Pie is more easily swipeable in at all times, but
a user might instead prefer being able to quickly swipe in from the
right or left hand side in portrait which becomes cumbersome when the
device is turned to landscape as the user is then forced to reach to
the middle of the long side of the screen which becomes located at
the bottom making it harder to access Pie.

The PIE_ALWAYS_AT_RIGHT value should be used to solve this issue
by forcing Pie to relocate itself to the right hand side just
like with the stock navigation bar.

Change-Id: I9b4dd3cee3fa7d85194389640dcedc02fa522530

Conflicts:
	packages/SystemUI/res/values-sw600dp/config.xml