frameworks/base: creating sperate files for ANRs traces
If an ANR comes just after the framework reboot, the traces for the
framework reboot will get over-written with the ANR traces. So, We
are missing the traces for the framework reboot which makes debugging
the framework reboot difficult.
Even for multiple ANRs in different apps, we will get the traces for
the last ANR.
Most of the time traces show that the call is stuck in some native call
and we do not have any info about the native call stack becuase we do
not collect the tombstones for teh ANR.
This patch will create one trace file for each application in case of ANR
and the filename would be traces_<app_name>.txt and the traces for
framework reboot will be stored in traces.txt
Change-Id: I5b78ce62b26f25d3dc14097e2988ba12c2d77826

Signed-off-by: houst0nn <houstonn@gmail.com>