WindowManager: Toast message is not visible in some scenarios
Issue: Toast does not get displayed if it is displayed from an
Activity which runs in a process different from the process of the
package to which the activity belongs.

Fix: While adding a window a check is done to ensure that the uid
of the activity matches the uid of the package to which the activity
belongs. This change will bypass the above check for Toast message.

CRs-Fixed: 720701

Change-Id: I29c5d30a8f11c2d132066c77bd038abc444cedb4
Signed-off-by: houst0nn <houstonn@gmail.com>