OnTheSpot: [1/2] Rework the OTS framework
Drop the animations and the graphics and everything so that the dialog
just looks like it was part of the system. Ask Arz more about this
change if you don't believe in it. Additionally, make the "ask later"
option a simple "just once" to avoid confusion about whether the button
will or will not make the UI feature appear. Along the way make the
whole android.util.SettingConfirmationHelper ordeal more adaptable to
all kinds of possible situations and move code around here and there.

Rework of Change-Id: I725a8164c13c5381ddbc15fde29902a258c1ecc4

To be merged with Change-Id:
  packages/apps/Settings: Ia1742fc3d7dfa5af9e7f29f68dc8ec76eea5a98d

Change-Id: I4cc5b89c7fc96d8dcc1cd65e2a1e03aff8c8af39

Conflicts:
	core/java/android/util/SettingConfirmationHelper.java
	core/res/res/values/pa_strings.xml