Missing unregistering PhoneStateListener doesn't cause heap leakage
First, implement Stub with static-nested class. By doing this, if the
PhoneStateListener's object is of no use, it can be collected as garbage
regardless of binder connection of Stub.
Second, if there remains Stub that is connected with TelephonyRegistry
after garbage collection of the PhoneStateListener's object, unregister
it internally, and does not cause heap leakage.

Change-Id: Ib06e4d25f1190aa97da0f31ac323e43c44a0ad9d
Signed-off-by: Dooyoung Hwang <dooyoung.hwang@lge.com>

Conflicts:
	telephony/java/android/telephony/PhoneStateListener.java