Screen Pinning: Show correct string for on screen nav
mPolicy.hasNavigationBar() will always return false because
a new PhoneWindowManager was created just for this one method
but the new PhoneWindowManager instance does not load the the
boolean from resources that define whether the device has on
screen navigation bar.  This patch removes the unnecessary instance
of a WindowManager and simply loads in the boolean when constructed.

Change-Id: I09c2ef661bff272cb4f7ca43bac0e45f4b20a4d4
TICKET: CYNGNOS-1111