SystemUI: handle when brightness slider is hidden
With this patch, the quick settings area should automatically expand and
make room for the tile trashcan/instruction area when the user has the
brightness slider hidden.

Also adds a scrollview to the QS Settings page to allow the list to grow
without impacting the height of the qs panel.

Ref: CYNGNOS-1644

Change-Id: I03f552ec28a978c4cad1e7e7a52648ec3d311a07
Signed-off-by: Roman Birg <roman@cyngn.com>

Patchset: 5
http://review.cyanogenmod.org/#/c/129044/

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/phone/NotificationPanelView.java