DreamManager: only notify power manager if lid state changes
In the case where the device is in landscape mode with a lockscreen
and the user presses the power button, PhoneWindowManager will update
it's configuration and eventually update the lid state.  If the lid
state is LID_OPEN, the doze override is reset in power manager.  This
causes the requested screen state to be STATE_UNKNOWN.  If the requested
screen state is STATE_UNKNOWN while in doze mode, the screen state will
be set to doze.  This causes the display to be left on in this case.

OPO-439 CYNGNOS-1701
Change-Id: Ib1cda5c7c30fc20f1a5c532d6b70478a46b22121

Patchset: 1
http://review.cyanogenmod.org/#/c/133050/