GlobalActions: Always use setZenMode
When set to silent mode, the user cannot change the ringer mode to
vibrate or ring.  Notification manager listens for changes to the
zen mode setting and will switch it back if there is a mismatch.
Using INotificationManager.setZenMode takes care of changing the mode
and keeping track of the updated change.

Change-Id: I5139a9a8cde4ca59f64bd21cf7c184d4745a7d59
TICKET: CYNGNOS-2417 BACON-4658