SystemUI: fix some qs tile race conditions
Whenever setTiles() was called, we assumed all the tile animations had
finished, but if the eager beaver had grabbed a tile and dropped it on the
remove target before waiting for all the animations to complete, a lot of
assumptions are broken and crashes happen. Mainly because tiles aren't removed
from pages and we try to add them to a different page (like if they
needed to be placed a page back).

Implemented cancel for ongoing animations, which we call whenever
setTiles comes in. So all tiles should be in their proper state before
we try to cache/re-add them. Also cleaned up and documented setTiles().

Also, don't try to place the edit tile in the proper place, just add it
to the list if it's not present.

Ticket: CYNGNOS-2472

Change-Id: I5c066abbc16f1fe7173525ea6a8a8b39460461ae
Signed-off-by: Roman Birg <roman@cyngn.com>

Conflicts:
	packages/SystemUI/src/com/android/systemui/qs/QSDragPanel.java