Apps on adopted storage should stay there during update
An update to an app on adopted storage via a PackageInstaller.Session
(eg gplay) will result in the app silently ending up back on internal
storage if the (new) manifest sets INSTALL_LOCATION_INTERNAL_ONLY.
However, the app data is not moved with it so the end result is that
data looks like it's been wiped.

Ideally, the app and data would be moved to internal.  Next best would
be that the user is informed that they need to uninstall and reinstall.

For now, persist the app in the adopted storage location so that
the user doesn't lose app data.

Change-Id: I1f178ea361a875c6df6b3b20a2e44071124c39c4

Patchset: 1
http://review.cyanogenmod.org/#/c/143846/